+++ migration/v3/migrate.sql
@@ -1,0 +1,2 @@
+UPDATE Images SET `filePath` = REPLACE (`filePath`,'/home/ubuntu','file:///home/ubuntu') WHERE `filePath` LIKE '/home/ubuntu%';
+ALTER TABLE Messages MODIFY COLUMN type enum('greeting','orderConfirm','paymentConfirm','deliveryConfirm','orderSync', 'forgotPassword', 'contact');