+++ migration/changelog.md
@@ -32,10 +32,22 @@
 
 * 新增 Contacts Table 用於聯繫表單
 
+
 ### 新增 Event 相關 Table
 
 * 新增 EventOrders
 * 新增 Events
 * Post type 新增 'internal-event','external-event' enum
-* Post url 清空目前沒有地方有用到 
+* Post url 清空目前沒有地方有用到
 * Post url 改名為 alias 放 url 別名
+
+## v3 
+
+### 紀錄方式改變
+
+* 更新 Image 的格式
+原本 `local` 的紀錄格式由 `/home/ubuntu/cargocms/.tmp/public/uploads/<<filename>>`，更新為 `file:///home/ubuntu/cargocms/.tmp/public/uploads/<<filename>>`
+
+### Message 欄位變動
+
+*  type 新增 contact enum